+++ contracts/staking/interfaces/IRewardCalculator.sol
@@ -7,7 +7,7 @@ pragma solidity 0.8.25;
 
 interface IRewardCalculator {
     function calculateReward(
-        uint64 stakeAmount,
+        uint256 stakeAmount,
         uint64 startTime,
         uint64 endTime,
         uint256 initialSupply,
